% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{m}
\alias{m}
\title{Assemble string with interpolated values}
\usage{
m(..., asis = getOption("m_asis", FALSE))
}
\arguments{
\item{...}{Any number of strings (to be concatenated).}

\item{asis}{If TRUE, output is generated with \code{cat}; in this case, the function must be called inside a chunk with \code{results='asis'}. If FALSE, output is generated with \code{knitr::asis_output}; in this case, \code{results='asis'} is not necessary.}
}
\value{
A character vector after interpolation.
}
\description{
Expressions enclosed by \verb{<<} and \verb{>>} will be evaluated as Python (or optionally R) code. Long strings are broken by line and concatenated together. Leading whitespace and blank lines from the first and last lines are automatically trimmed.
}
\author{
fnaufel
}
