% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{sympy_transformer}
\alias{sympy_transformer}
\title{Custom transformer}
\usage{
sympy_transformer(text, envir)
}
\arguments{
\item{text}{Contents between one \verb{<<...>>} pair.}

\item{envir}{Not used.}
}
\value{
Character vector after evaluation.
}
\description{
Custom transformer
}
\author{
fnaufel
}
